updata NLP model
调整了NLP模块的结构
