+++ NLP/service/gpt/azure_gpt.go
@@ -1,0 +1,7 @@
+package gpt
+
+import "GoTuber/MESSAGE/model"
+
+func GenerateTextByAzureOpenAI(msg *model.Msg) {
+	//TODO:完成这个
+}
